@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2023 karamem0
+// Copyright (c) 2022-2024 karamem0
 //
 // This software is released under the MIT License.
 //
@@ -57,7 +57,7 @@ namespace Karamem0.Commistant.Functions.Commands.Tests
                 .Returns(new ResourceResponse());
             var qrCodeService = Substitute.For<IQrCodeService>();
             _ = qrCodeService.CreateAsync("https://www.example.com")
-                .Returns(Array.Empty<byte>());
+                .Returns([]);
             var logger = Substitute.For<ILogger<StartMeetingCommand>>();
             logger.StartMeetingMessageNotifying(conversationReference, conversationProperty);
             logger.StartMeetingMessageNotified(conversationReference, conversationProperty);
@@ -112,7 +112,7 @@ namespace Karamem0.Commistant.Functions.Commands.Tests
                 .Returns(new ResourceResponse());
             var qrCodeService = Substitute.For<IQrCodeService>();
             _ = qrCodeService.CreateAsync("https://www.example.com")
-                .Returns(Array.Empty<byte>());
+                .Returns([]);
             var logger = Substitute.For<ILogger<StartMeetingCommand>>();
             logger.StartMeetingMessageNotifying(conversationReference, conversationProperty);
             logger.StartMeetingMessageNotified(conversationReference, conversationProperty);
@@ -167,7 +167,7 @@ namespace Karamem0.Commistant.Functions.Commands.Tests
                 .Returns(new ResourceResponse());
             var qrCodeService = Substitute.For<IQrCodeService>();
             _ = qrCodeService.CreateAsync("https://www.example.com")
-                .Returns(Array.Empty<byte>());
+                .Returns([]);
             var logger = Substitute.For<ILogger<StartMeetingCommand>>();
             logger.StartMeetingMessageNotifying(conversationReference, conversationProperty);
             logger.StartMeetingMessageNotified(conversationReference, conversationProperty);
@@ -222,7 +222,7 @@ namespace Karamem0.Commistant.Functions.Commands.Tests
                 .Returns(new ResourceResponse());
             var qrCodeService = Substitute.For<IQrCodeService>();
             _ = qrCodeService.CreateAsync("https://www.example.com")
-                .Returns(Array.Empty<byte>());
+                .Returns([]);
             var logger = Substitute.For<ILogger<StartMeetingCommand>>();
             logger.StartMeetingMessageNotifying(conversationReference, conversationProperty);
             logger.StartMeetingMessageNotified(conversationReference, conversationProperty);
@@ -277,7 +277,7 @@ namespace Karamem0.Commistant.Functions.Commands.Tests
                 .Returns(new ResourceResponse());
             var qrCodeService = Substitute.For<IQrCodeService>();
             _ = qrCodeService.CreateAsync("https://www.example.com")
-                .Returns(Array.Empty<byte>());
+                .Returns([]);
             var logger = Substitute.For<ILogger<StartMeetingCommand>>();
             logger.StartMeetingMessageNotifying(conversationReference, conversationProperty);
             logger.StartMeetingMessageNotified(conversationReference, conversationProperty);
